requirements.txt
firecrawl-py==1.5.0
supabase==2.10.0
openai==1.55.3
python-dotenv==1.0.1
asyncio==3.4.3
aiohttp==3.10.10
python-dateutil==2.9.0.post0
beautifulsoup4==4.12.3
lxml==5.3.0
schedule==1.2.2